requirements.txt
beautifultable==0.7.0
blis==0.4.1
blosc==1.8.1
catalogue==1.0.0
certifi==2021.5.30
chardet==4.0.0
click==6.7
cycler==0.11.0
cymem==2.0.3
dataclasses==0.7
docker==3.6.0
docker-pycreds==0.4.0
easydict==1.9
editdistance==0.6.2
einops==0.4.1
en-core-web-sm==2.2.5
en-vectors-web-lg==2.1.0
evalai==1.3.14
fastdtw==0.3.4
filelock==3.0.12
ftfy==6.0.3
huggingface-hub==0.0.12
idna==2.10
imagesize==1.4.1
importlib-metadata==1.6.0
ipdb==0.12.2
joblib==1.1.1
jsonpatch==1.26
jsonpointer==2.0
kiwisolver==1.3.1
lxml==4.6.2
matplotlib==3.3.4
mkl-fft==1.3.0
mkl-random==1.1.1
mkl-service==2.3.0
multimatch==0.1.0
multimatch-gaze==0.1.3
murmurhash==1.0.2
numpy @ file:///tmp/build/80754af9/numpy_and_numpy_base_1603487797006/work
olefile==0.46
opencv-python==4.7.0.68
packaging==21.3
pandas==1.1.5
Pillow==8.4.0
plac==1.1.3
preshed==3.0.2
pyparsing==3.0.9
python-dateutil==2.7.3
pytorch-transformers==1.2.0
pytz==2023.3
PyYAML==6.0
regex==2020.6.8
requests==2.25.1
ruamel-yaml-conda @ file:///tmp/build/80754af9/ruamel_yaml_1616016274869/work
sacremoses==0.0.43
scipy==1.2.1
seaborn==0.11.2
sentencepiece==0.1.91
six @ file:///tmp/build/80754af9/six_1644875935023/work
spacy==2.2.4
srsly==1.0.2
tensorboardX==2.0
thinc==7.4.0
timm==0.4.9
tokenizers==0.10.3
torch==1.7.1
torchaudio==0.7.0a0+a853dff
torchfile==0.1.0
torchnet==0.0.4
torchvision==0.8.2
tqdm==4.62.3
transformers==4.8.1
typing_extensions @ file:///opt/conda/conda-bld/typing_extensions_1647553014482/work
urllib3==1.26.14
validators==0.12.6
visdom==0.1.8.9
wasabi==0.6.0
websocket-client==0.57.0
zipp==3.6.0
